model: adjust data type

+++ src/main/java/restful/api/eztrain/model/RegisterRouteRequest.java
@@ -22,6 +22,6 @@ public class RegisterRouteRequest {
     private Double tripDistance;
 
     @NotNull
-    private Integer tripDuration;
+    private Double tripDuration;
 
 }
